Move the per-op opcode handlers to vm_handlers.go
Pure file reorganization: the opcode-ordered leaf handlers (loads and
stores, scalar comparisons, division, float and truncation ops, SIMD)
move out of vm.go, which keeps the VM machinery, the dispatch switch,
control flow, and the handlers that touch store structures. Verified
byte-identical per-function code via go tool nm -size and a flat 17/17
benchmark comparison against the previous commit.
